Fix start of wifi tethering in a0
This patch removes the start of dnsmasq when property for a0 is set.
Otherwise the start of dnsmasq would fail and causes the tether service
to restart in an infinit loop. As we run a dnsmasq instance for virtual
network anyway we have to use this also for tethering.

Jira -> TRUSTME-1586

Change-Id: I86fb6b90ad768741d32c4e926d5e17d8e29533c0