Fix IpSecManager rekeying for incoming transforms
The IpSecManager documentation describes a rekeying behavior
where incoming transforms remain active until they are explicitly
deallocated by close().

However, the current behavior differs from the documented one:
If a transform A is active and a transform B is applied, incoming
IPSec packets using key, SPI, etc of Transform A are dropped
instantaneously although transform A was not deallocated using close().

The reason for this is that the XfrmController explicitly specifies
the SPI in the IPSec policy. When transform B is activated, the SPI
of the underlying IPSec policy is updated to match the SPI
used in the new transform:

emulator64_x86_64:/ # ip xfrm policy
src 0.0.0.0/0 dst 0.0.0.0/0
        socket in priority 0 flag localok
        tmpl src 192.168.11.2 dst 10.0.2.15
                proto esp spi 0x000003e9 reqid 10105 mode transport
emulator64_x86_64:/ # ip xfrm state
src 192.168.11.2 dst 10.0.2.15
        proto esp spi 0x000003e9 reqid 10105 mode transport
        replay-window 32
        auth-trunc hmac(sha256) 0x0000000000000000000000000000000000000000000000000000000000000000 128
        enc cbc(aes) 0x0000000000000000000000000000000000000000000000000000000000000000
        anti-replay context: seq 0x0, oseq 0x0, bitmap 0x00000000
        sel src 0.0.0.0/0 dst 0.0.0.0/0
src 192.168.11.2 dst 10.0.2.15
        proto esp spi 0x000003e8 reqid 10105 mode transport
        replay-window 32
        auth-trunc hmac(sha256) 0x0000000000000000000000000000000000000000000000000000000000000000 128
        enc cbc(aes) 0x0000000000000000000000000000000000000000000000000000000000000000
        anti-replay context: seq 0x0, oseq 0x0, bitmap 0x00000000
        sel src 0.0.0.0/0 dst 0.0.0.0/0

This causes the kernel to reject packets with the SPI of transform A
although the SA for transform a still exists:

In order to achieve the documented behavior, the SPI must not
be specified in policies created on behalf of incoming IPSecManager
transforms. This way, both SAs created on behalf of transform A and B
match incoming packets achieving the decryption of in-flight packets
created using transform A's SPI.

This commit therefore adapts the XfrmController disabling the SPI
in incoming policies created on behalf of IpSecManager transforms.